backend/requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-autoslug==1.9.9
django-cors==0.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
flake8==7.1.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.40
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-decouple==3.8
regex==2024.5.15
requests==2.32.3
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
